Add Statistics interface and implementations
These implementations are merge-able with their same type in order to easily create combined statistics.
As the design of the actual statistics service is yet to be decided, the methods provided by implementations of Statistics might be subject to change. However, the Statistics interface itself should be safe to use and the implementations are functional for now.